Implement choose_context_to_advance_promise
This also makes the system totally lazy.

As far as I understand it, a context C can advance a promise P iff
- C's answer can fulfill P,
- C's final workspace can fulfill P or
- one of C's promisees (which are contexts) can advance the P.

The implementation follows this definition, so it should be correct… but
of course I might be fooling myself.

The implementation uses recursion without performance enhancements, such
as memoization. For now this should be fine. We can improve it later if
necessary.