Add test_basic
It would take a while to put together a proper test suite, but for now I
at least want to know when my changes break the most basic scenario(s).
To this end add test_basic.

The test is hard to read and tightly coupled with the implementation.
Writing test support code to get rid of the coupling would be tedious,
but might become necessary.