REFACTOR RootQuestionSession: Tidy up variables
Remove an instance variable that wasn't used. Rename
"final_answer_promise" and "final_answer" to "root_answer_promise" and
"root_answer". "final" sounds as if after resolving the promise nothing
needs to be done. However, if the answer contains pointers, those need
to be unlocked. And if they point to promises, they need to be resolved.
Besides, it's a RootQuestionSession, and a root question should have a
root answer.